refactor : graph 중복 제거
